+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [7.1.0] - 2021-09-16
+## [7.1.0]
+
+### Added
+
+-   Support for multiple access token signing keys: https://github.com/supertokens/supertokens-core/issues/305
+-   Supporting CDI 2.9
+-   Header `"content-type": "application/json; charset=utf-8"` when querying the core
 
 ### Added
 
@@ -15,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 -   Unnecessary serverless optimisation
+
+# [7.0.2] - 2021-09-16
+
+### Fixes
+
+-   fixes issue #180 where user was required to have all frameworks as depedencies or else use `compilerOptions.skipLibCheck` in their tsconfig.json
+-   `verifySession` middleware ts issue fix for Hapi framework
+-   ts import for `verifySession` fix for koa
 
 ## [7.0.1] - 2021-09-10